refactor to split cli and lib
