+++ jello/__init__.py
@@ -1,0 +1,8 @@
+"""jello - query JSON at the command line with python syntax"""
+
+
+__version__ = '1.3.7'
+AUTHOR = 'Kelly Brazil'
+WEBSITE = 'https://github.com/kellyjonbrazil/jello'
+COPYRIGHT = '© 2020-2021 Kelly Brazil'
+LICENSE = 'MIT License'
