+++ jello/cli.py
@@ -4,7 +4,6 @@ import os
 import sys
 import platform
 import textwrap
-import json
 import signal
 import ast
 from pygments import highlight
@@ -12,14 +11,10 @@ from pygments.style import Style
 from pygments.token import (Name, Number, String, Keyword)
 from pygments.lexers import JsonLexer
 from pygments.formatters import Terminal256Formatter
-from jello.dotmap import DotMap
+import jello
+from jello.lib import load_json, create_schema, create_json, pyquery
+from jello.options import opts, JelloTheme
 
-
-__version__ = '1.3.7'
-AUTHOR = 'Kelly Brazil'
-WEBSITE = 'https://github.com/kellyjonbrazil/jello'
-COPYRIGHT = '© 2020-2021 Kelly Brazil'
-LICENSE = 'MIT License'
 
 color_map = {
     'black': ('ansiblack', '\33[30m'),
@@ -39,32 +34,6 @@ color_map = {
     'brightcyan': ('ansibrightcyan', '\33[96m'),
     'white': ('ansiwhite', '\33[97m'),
 }
-
-
-class opts:
-    initialize = None
-    version_info = None
-    helpme = None
-    compact = None
-    nulls = None
-    raw = None
-    lines = None
-    mono = None
-    schema = None
-    keyname_color = None
-    keyword_color = None
-    number_color = None
-    string_color = None
-    arrayid_color = None
-    arraybracket_color = None
-
-
-schema_list = []
-
-
-class JelloTheme:
-    """this class will contain the colors dictionary generated from set_env_colors()"""
-    pass
 
 
 def set_env_colors():
@@ -168,262 +137,6 @@ def helptext():
     sys.exit()
 
 
-def create_schema(src, path=''):
-    """
-    Creates a grep-able schema representation of the JSON.
-
-    This function is recursive, so output is stored within the schema_list list. Make sure to
-    initialize schema_list to a blank list and set colors by calling set_env_colors() before
-    calling this function.
-    """
-    if not opts.mono:
-        CEND = '\33[0m'
-        CBOLD = '\33[1m'
-        CKEYNAME = f'{JelloTheme.colors["key_name"][1]}'
-        CKEYWORD = f'{JelloTheme.colors["keyword"][1]}'
-        CNUMBER = f'{JelloTheme.colors["number"][1]}'
-        CSTRING = f'{JelloTheme.colors["string"][1]}'
-        CARRAYID = f'{JelloTheme.colors["array_id"][1]}'
-        CARRAYBRACKET = f'{JelloTheme.colors["array_bracket"][1]}'
-
-    else:
-        CEND = ''
-        CBOLD = ''
-        CKEYNAME = ''
-        CKEYWORD = ''
-        CNUMBER = ''
-        CSTRING = ''
-        CARRAYID = ''
-        CARRAYBRACKET = ''
-
-    if isinstance(src, list) and path == '':
-        for i, item in enumerate(src):
-            create_schema(item, path=f'.{CARRAYBRACKET}[{CEND}{CARRAYID}{i}{CEND}{CARRAYBRACKET}]{CEND}')
-
-    elif isinstance(src, list):
-        for i, item in enumerate(src):
-            create_schema(item, path=f'{path}.{CBOLD}{CKEYNAME}{src}{CEND}{CARRAYBRACKET}[{CEND}{CARRAYID}{i}{CEND}{CARRAYBRACKET}]{CEND}')
-
-    elif isinstance(src, dict):
-        for k, v in src.items():
-            if isinstance(v, list):
-                for i, item in enumerate(v):
-                    create_schema(item, path=f'{path}.{CBOLD}{CKEYNAME}{k}{CEND}{CARRAYBRACKET}[{CEND}{CARRAYID}{i}{CEND}{CARRAYBRACKET}]{CEND}')
-
-            elif isinstance(v, dict):
-                if not opts.mono:
-                    k = f'{CBOLD}{CKEYNAME}{k}{CEND}'
-                create_schema(v, path=f'{path}.{k}')
-
-            else:
-                k = f'{CBOLD}{CKEYNAME}{k}{CEND}'
-                val = json.dumps(v, ensure_ascii=False)
-                if val == 'true' or val == 'false' or val == 'null':
-                    val = f'{CKEYWORD}{val}{CEND}'
-                elif val.replace('.', '', 1).isdigit():
-                    val = f'{CNUMBER}{val}{CEND}'
-                else:
-                    val = f'{CSTRING}{val}{CEND}'
-
-                schema_list.append(f'{path}.{k} = {val};')
-
-    else:
-        val = json.dumps(src, ensure_ascii=False)
-        if val == 'true' or val == 'false' or val == 'null':
-            val = f'{CKEYWORD}{val}{CEND}'
-        elif val.replace('.', '', 1).isdigit():
-            val = f'{CNUMBER}{val}{CEND}'
-        else:
-            val = f'{CSTRING}{val}{CEND}'
-
-        path = path or '.'
-
-        schema_list.append(f'{path} = {val};')
-
-
-def create_json(data):
-    separators = None
-    indent = 2
-
-    if opts.compact or opts.lines:
-        separators = (',', ':')
-        indent = None
-
-    if isinstance(data, dict):
-        return json.dumps(data, separators=separators, indent=indent, ensure_ascii=False)
-
-    if isinstance(data, list):
-        if not opts.lines:
-            return json.dumps(data, separators=separators, indent=indent, ensure_ascii=False)
-
-        # check if this list includes lists
-        list_includes_list = False
-        for item in data:
-            if isinstance(item, list):
-                list_includes_list = True
-                break
-
-        if opts.lines and list_includes_list:
-            raise ValueError('Cannot print list of lists as lines. Try normal JSON output.')
-
-        # print lines for a flat list
-        else:
-            flat_list = ''
-            for entry in data:
-                if entry is None:
-                    if opts.nulls:
-                        flat_list += 'null\n'
-                    else:
-                        flat_list += '\n'
-
-                elif isinstance(entry, (dict, bool, int, float)):
-                    flat_list += json.dumps(entry, separators=separators, ensure_ascii=False) + '\n'
-
-                elif isinstance(entry, str):
-                    # replace \n with \\n here so lines with newlines literally print the \n char
-                    entry = entry.replace('\n', '\\n')
-                    if opts.raw:
-                        flat_list += f'{entry}' + '\n'
-                    else:
-                        flat_list += f'"{entry}"' + '\n'
-
-            return flat_list.rstrip()
-
-    # naked single item return case
-    elif data is None:
-        if opts.nulls:
-            return 'null'
-        else:
-            return ''
-
-    elif isinstance(data, (bool, int, float)):
-        return json.dumps(data, ensure_ascii=False)
-
-    elif isinstance(data, str):
-        # replace \n with \\n here so lines with newlines literally print the \n char
-        data = data.replace('\n', '\\n')
-        if opts.raw:
-            return f'{data}'
-        else:
-            return f'"{data}"'
-
-    # only non-serializable types are left. Force an exception from json.dumps()
-    else:
-        json.dumps(data)
-        # this code should not run, but just in case something slips by above
-        raise TypeError(f'Object is not JSON serializable')
-
-def pyquery(data, query):
-    # if data is a list of dictionaries, then need to iterate through and convert all dictionaries to DotMap
-    if isinstance(data, list):
-        _ = [DotMap(i, _dynamic=False, _prevent_method_masking=True) if isinstance(i, dict)
-             else i for i in data]
-
-    elif isinstance(data, dict):
-        _ = DotMap(data, _dynamic=False, _prevent_method_masking=True)
-
-    else:
-        _ = data
-
-    jelloconf = ''
-    conf_file = ''
-
-    if opts.initialize:
-        if platform.system() == 'Windows':
-            conf_file = os.path.join(os.environ['APPDATA'], '.jelloconf.py')
-        else:
-            conf_file = os.path.join(os.environ["HOME"], '.jelloconf.py')
-
-        try:
-            with open(conf_file, 'r') as f:
-                jelloconf = f.read()
-        except FileNotFoundError:
-            raise FileNotFoundError(f'-i used and initialization file not found: {conf_file}')
-
-    query = jelloconf + query
-    output = None
-
-    # extract jello options from .jelloconf.py (compact, raw, lines, nulls, mono, and custom colors)
-    for expr in ast.parse(jelloconf).body:
-        if isinstance(expr, ast.Assign):
-            if expr.targets[0].id == 'compact':
-                opts.compact = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'raw':
-                opts.raw = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'lines':
-                opts.lines = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'nulls':
-                opts.nulls = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'mono':
-                opts.mono = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'schema':
-                opts.schema = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'keyname_color':
-                opts.keyname_color = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'keyword_color':
-                opts.keyword_color = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'number_color':
-                opts.number_color = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'string_color':
-                opts.string_color = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'arrayid_color':
-                opts.arrayid_color = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-            if expr.targets[0].id == 'arraybracket_color':
-                opts.arraybracket_color = eval(compile(ast.Expression(expr.value), '<string>', "eval"))
-
-            # validate the data in the initialization file
-            warn_options = False
-            warn_colors = False
-
-            for option in [opts.compact, opts.raw, opts.lines, opts.nulls, opts.mono, opts.schema]:
-                if not isinstance(option, bool):
-                    opts.compact = opts.raw = opts.lines = opts.nulls = opts.mono = opts.schema = None
-                    warn_options = True
-
-            for color_config in [opts.keyname_color, opts.keyword_color, opts.number_color,
-                                 opts.string_color, opts.arrayid_color, opts.arraybracket_color]:
-                valid_colors = ['black', 'red', 'green', 'yellow', 'blue', 'magenta', 'cyan', 'gray', 'brightblack', 'brightred',
-                                'brightgreen', 'brightyellow', 'brightblue', 'brightmagenta', 'brightcyan', 'white']
-                if color_config not in valid_colors and color_config is not None:
-                    opts.keyname_color = opts.keyword_color = opts.number_color = opts.string_color = opts.arrayid_color = opts.arraybracket_color = None
-                    warn_colors = True
-
-            if warn_options:
-                print(f'Jello:   Warning: Options must be set to True or False in {conf_file}\n         Unsetting all options.\n')
-
-            if warn_colors:
-                valid_colors_string = ', '.join(valid_colors)
-                print(f'Jello:   Warning: Colors must be set to one of: {valid_colors_string} in {conf_file}\n         Unsetting all colors.\n')
-
-    # run the query
-    block = ast.parse(query, mode='exec')
-    last = ast.Expression(block.body.pop().value)    # assumes last node is an expression
-    exec(compile(block, '<string>', mode='exec'))
-    output = eval(compile(last, '<string>', mode='eval'))
-
-    # convert output back to normal dict
-    if isinstance(output, list):
-        output = [i.toDict() if isinstance(i, DotMap) else i for i in output]
-
-    elif isinstance(output, DotMap):
-        output = output.toDict()
-
-    # if DotMap returns a bound function then we know it was a reserved attribute name
-    if hasattr(output, '__self__'):
-        raise ValueError('Reserved key name. Use bracket notation to access this key.')
-
-    return output
-
-
-def load_json(data):
-    try:
-        json_dict = json.loads(data)
-    except Exception:
-        # if json.loads fails, assume the data is json lines and parse
-        json_dict = [json.loads(i) for i in data.splitlines()]
-
-    return json_dict
-
 def format_exception(e=None, list_dict_data='', query='', response='', output=''):
     query = str(query).replace('\n', '; ')
     e_text = ''
@@ -433,7 +146,7 @@ def format_exception(e=None, list_dict_data='', query='', response='', output=''
 
     if len(str(list_dict_data)) > 70:
         list_dict_data = str(list_dict_data)[:34] + ' ... ' + str(list_dict_data)[-34:]
-   
+
     if len(str(query)) > 70:
         query = str(query)[:34] + ' ... ' + str(query)[-34:]
 
@@ -497,11 +210,11 @@ def main(data=None, query='_'):
 
     if opts.version_info:
         print(textwrap.dedent(f'''\
-            jello:   Version: {__version__}
-                     Author: {AUTHOR}
-                     Website: {WEBSITE}
-                     Copyright: {COPYRIGHT}
-                     License: {LICENSE}
+            jello:   Version: {jello.__version__}
+                     Author: {jello.AUTHOR}
+                     Website: {jello.WEBSITE}
+                     Copyright: {jello.COPYRIGHT}
+                     License: {jello.LICENSE}
         '''))
         sys.exit()
 
@@ -527,7 +240,7 @@ def main(data=None, query='_'):
         try:
             response = pyquery(list_dict_data, query)
 
-        except Exception as e:    
+        except Exception as e:
             e, e_text, list_dict_data, query, response, output = format_exception(e,
                                                                                   list_dict_data,
                                                                                   query)
@@ -559,7 +272,7 @@ def main(data=None, query='_'):
                 if not sys.stdout.isatty():
                     opts.mono = True
                 create_schema(response)
-                print('\n'.join(schema_list))
+                print('\n'.join(jello.lib.schema_list))
                 sys.exit()
             else:
                 output = create_json(response)
