+++ jello/options.py
@@ -1,0 +1,24 @@
+"""jello - query JSON at the command line with python syntax"""
+
+
+class opts:
+    initialize = None
+    version_info = None
+    helpme = None
+    compact = None
+    nulls = None
+    raw = None
+    lines = None
+    mono = None
+    schema = None
+    keyname_color = None
+    keyword_color = None
+    number_color = None
+    string_color = None
+    arrayid_color = None
+    arraybracket_color = None
+
+
+class JelloTheme:
+    """this class will contain the colors dictionary generated from set_env_colors()"""
+    pass
